chore: moves Tranding and Predict to tests (#25219)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
Following https://github.com/MetaMask/metamask-mobile/pull/24313 we're
looking to centralize all tools and test resources in one place.
This PR moves page objects, selectors and spec files for `Trending` and
`Predict` to `/tests`.

Previous related PRs:
- https://github.com/MetaMask/metamask-mobile/pull/24988
- https://github.com/MetaMask/metamask-mobile/pull/24313
- https://github.com/MetaMask/metamask-mobile/pull/25031
- https://github.com/MetaMask/metamask-mobile/pull/25095
- https://github.com/MetaMask/metamask-mobile/pull/25167
- https://github.com/MetaMask/metamask-mobile/pull/25198


### ⚠️ **Important note:**
`e2e-split-tags-shards.mjs` was changed to start finding for a given tag
starting from the root. This affects drastically the performance of
searching for specs. Search times go from ~0.7s to ~2.5s. This is needed
for a phased out transition from `/e2e/specs` to `/tests/specs`


<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry:

## **Related issues**

Fixes:

## **Manual testing steps**

```gherkin
Feature: my feature name

  Scenario: user [verb for user action]
    Given [describe expected initial app state]

    When user [verb for user action]
    Then [describe expected outcome]
```

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I've included tests if applicable
- [ ] I've documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I've applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> Centralizes E2E assets for Trending and Predict under `tests/`,
aligning page objects, locators, and smoke/spec files.
> 
> - Moved `Trending` and `Predict` page objects to
`tests/page-objects/**` and selectors to `tests/locators/**`; updated
e2e specs/imports accordingly
> - Added/updated smoke tests under `tests/smoke/predict/**` and shared
helpers
> - Updated `e2e/jest.e2e.config.js` `testMatch` to include
`tests/**/*.spec.{js,ts}`
> - Changed `e2e-split-tags-shards.mjs` default `BASE_DIR` to repo root
(`./`) to find tagged specs across both `e2e/` and `tests/`
> - Adjusted `TabBarComponent` and `viewHelper` to reference new
`TrendingView` paths
> 
> <sup>Written by [Cursor
Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit
1a4f28075b9ced57e9d28da2e09278fd327f4170. This will update automatically
on new commits. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->
