+++ e2e/jest.e2e.config.js
@@ -7,7 +7,10 @@ require('dotenv').config({ path: '.e2e.env' });
 
 module.exports = {
   rootDir: '..',
-  testMatch: ['<rootDir>/e2e/specs/**/*.spec.{js,ts}'],
+  testMatch: [
+    '<rootDir>/e2e/specs/**/*.spec.{js,ts}',
+    '<rootDir>/tests/**/*.spec.{js,ts}',
+  ],
   testTimeout: 300000,
   maxWorkers: 1,
   clearMocks: true,
