+++ tests/locators/Trending/TrendingView.selectors.ts
