+++ tests/page-objects/Predict/PredictMarketList.ts
@@ -1,4 +1,4 @@
-import { Gestures, Matchers } from '../../../tests/framework';
+import { Gestures, Matchers } from '../../framework';
 import {
   PredictMarketListSelectorsIDs,
   getPredictMarketListSelector,
